parser: split into {item,module}.rs
